Rename edit_prediction -> edit_prediction_types
