+++ Cargo.toml
@@ -53,7 +53,7 @@ members = [
     "crates/denoise",
     "crates/diagnostics",
     "crates/docs_preprocessor",
-    "crates/edit_prediction",
+    "crates/edit_prediction_types",
     "crates/edit_prediction_button",
     "crates/edit_prediction_context",
     "crates/zeta2_tools",
@@ -313,7 +313,7 @@ http_client = { path = "crates/http_client" }
 http_client_tls = { path = "crates/http_client_tls" }
 icons = { path = "crates/icons" }
 image_viewer = { path = "crates/image_viewer" }
-edit_prediction = { path = "crates/edit_prediction" }
+edit_prediction_types = { path = "crates/edit_prediction_types" }
 edit_prediction_button = { path = "crates/edit_prediction_button" }
 edit_prediction_context = { path = "crates/edit_prediction_context" }
 zeta2_tools = { path = "crates/zeta2_tools" }
