+++ crates/codestral/Cargo.toml
@@ -10,7 +10,7 @@ path = "src/codestral.rs"
 
 [dependencies]
 anyhow.workspace = true
-edit_prediction.workspace = true
+edit_prediction_types.workspace = true
 edit_prediction_context.workspace = true
 futures.workspace = true
 gpui.workspace = true
