+++ crates/copilot/Cargo.toml
@@ -33,7 +33,7 @@ fs.workspace = true
 futures.workspace = true
 gpui.workspace = true
 http_client.workspace = true
-edit_prediction.workspace = true
+edit_prediction_types.workspace = true
 language.workspace = true
 log.workspace = true
 lsp.workspace = true
