+++ crates/edit_prediction_button/Cargo.toml
@@ -18,7 +18,7 @@ client.workspace = true
 cloud_llm_client.workspace = true
 codestral.workspace = true
 copilot.workspace = true
-edit_prediction.workspace = true
+edit_prediction_types.workspace = true
 editor.workspace = true
 feature_flags.workspace = true
 fs.workspace = true
