+++ crates/supermaven/Cargo.toml
@@ -16,7 +16,7 @@ doctest = false
 anyhow.workspace = true
 client.workspace = true
 collections.workspace = true
-edit_prediction.workspace = true
+edit_prediction_types.workspace = true
 futures.workspace = true
 gpui.workspace = true
 language.workspace = true
