+++ crates/zeta/Cargo.toml
@@ -28,7 +28,7 @@ command_palette_hooks.workspace = true
 copilot.workspace = true
 credentials_provider.workspace = true
 db.workspace = true
-edit_prediction.workspace = true
+edit_prediction_types.workspace = true
 edit_prediction_context.workspace = true
 editor.workspace = true
 feature_flags.workspace = true
